backend/requirements.txt
Flask
flask-cors
Pillow

